pyproject.toml
[project]
name = "pose-deploy-gate"
version = "0.2.0"
description = "Deployment-focused validation tooling for body keypoint model comparison."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # import sorting
    "B",   # bugbear
    "UP",  # pyupgrade
]

ignore = [
    "E501" # line too long handled by formatter
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/pose_deploy_gate"]
